+++ GringottsAPI/Models/AddCustomerInputModel.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public class AddCustomerInputModel
     {
-        public int CustomerNumber { get; set; }
+        public long CustomerNumber { get; set; }
         public string FirstName { get; set; }
         public string MiddleName { get; set; }
         public string LastName { get; set; }
